#195 split ClientAction to detail actions
