+++ src/static/app/client/client-content/ex-schema-policy-check.tag
@@ -316,7 +316,7 @@
         if (self.checkModal.status === 'Check') {
           self.refs.checkModal.hide()
         }
-        ApiFactory.clientOperation(self.opts.projectName).then((response) => {
+        ApiFactory.clientPropbase(self.opts.projectName).then((response) => {
           self.client = response
           self.updateLatestResult(self.client)
           self.update()
