+++ src/static/app/common/DbfluteTask.js
@@ -7,7 +7,7 @@ export default class DbfluteTask {
     return ApiFactory.task(projectName, task).then((response) => {
       const message = response.success ? 'success' : 'failure'
       callback(message)
-      ApiFactory.clientOperation(projectName).then((response) => {
+      ApiFactory.clientPropbase(projectName).then((response) => {
         self.update({
           client: response
         })
